Slice judgment view tests into their own file
The number of tests we have for judgment view behaviours should be a lot higher, so pre-emptively put tests for these views in their own file.
